refactor: extract mail list hooks
